Imported dto's from monolith. Minor refactor. Bumped ver 1.0.3
