+++ common-data-lib/pom.xml
@@ -6,7 +6,7 @@
 
   <groupId>com.sainnt.netshop</groupId>
   <artifactId>common-data-lib</artifactId>
-  <version>1.0.2</version>
+  <version>1.0.3</version>
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <kotlin.version>1.6.21</kotlin.version>
