+++ common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/crm/ProductDto.kt
@@ -1,0 +1,8 @@
+package com.sainnt.netshop.common.dto.crm
+
+data class ProductDto(
+    val id: Long,
+    val name:String,
+    val catalogId: Long,
+    val price: Long
+)
