+++ common-data-lib/src/main/kotlin/com/sainnt/netshop/common/dto/request/ProductCreateDto.kt
@@ -1,0 +1,8 @@
+package com.sainnt.netshop.common.dto.request
+
+data class ProductCreateDto(
+    val name: String,
+    val catalogId: Long,
+    val price: Long,
+    val description: String? = null
+)
